requirements.txt
altair==5.5.0
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-mysql==4.14.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gensim==4.3.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.27.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
narwhals==1.19.1
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pbr==6.1.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
protobuf==5.29.2
punk==3.0.0
punktdict==0.11
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.5.0
stevedore==5.3.0
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.3
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.26.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
whitenoise==6.4.0
wordcloud==1.9.4
wrapt==1.17.0
